+++ ossm-sauce-firmware/src/Ossm.h
@@ -1,5 +1,5 @@
 #include "MotorMovement.h"
-#include "CommandType.h"
+#include "Common.h"
 
 void initializeMoveQueue();
 void updateState();
@@ -7,16 +7,16 @@ void updateState();
 // Command Controls
 void enqueueMove(StrokeCommand movement);
 void initiateLoop(StrokeCommand loopPushInput, StrokeCommand loopPullInput);
-void moveToPosition(uint32_t positionInput);
+void moveToPosition(int positionInput);
 void initiateVibrate(Vibration vibrationInput);
 void initiateSmoothMove(StrokeCommand smoothMoveInput);
 void play(MovementMode movementModeInput);
 void play(MovementMode movementModeInput, unsigned long playTimeMsInput);
 void pauseNow();
 void resetNow();
-void initiateHoming(uint32_t positionInput);
+void initiateHoming(int positionInput);
 void setSpeedLimit(int speedLimit);
 void setGlobalAcceleration(int acceleration);
 void setRangeLimit(short rangeLimitInput, byte selectedRange);
-void setHomingSpeed(uint32_t homingSpeedHzInput);
+void setHomingSpeed(unsigned long homingSpeedHzInput);
 void setHomingTrigger(float homingTriggerInput);
